+++ src/client/components/Navigation/MobileMenu/MobileMenu.js
@@ -2,25 +2,19 @@ import React, { useRef } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { get, isUndefined, size, filter, includes } from 'lodash';
 import { FormattedMessage, injectIntl } from 'react-intl';
-import { AutoComplete, Button, Input, Menu } from 'antd';
+import { AutoComplete, Button, Input } from 'antd';
 import classNames from 'classnames';
 import ModalBroker from '../../../../investarena/components/Modals/ModalBroker';
-import BTooltip from '../../BTooltip';
-import PopoverContainer from '../../Popover';
-import Notifications from '../Notifications/Notifications';
-import { PARSED_NOTIFICATIONS } from '../../../../common/constants/notifications';
 import { disconnectBroker } from '../../../../investarena/redux/actions/brokersActions';
 import { logout } from '../../../auth/authActions';
 import { getIsBeaxyUser } from '../../../user/usersHelper';
 import BrokerBalance from '../BrokerBalance/BrokerBalance';
 import Avatar from '../../Avatar';
 import SignUp from '../../Sidebar/SignUp';
-import HotNews from '../HotNews';
-import { getModalIsOpenState } from '../../../../investarena/redux/selectors/modalsSelectors';
-import './MobileMenu.less';
 import LoggedMenuMobile from './LoggedMenuMobile/LoggedMenuMobile';
+import './MobileMenu.less';
+import MenuButtons from './MenuButtons/MenuButtons';
 
 const MobileMenu = props => {
   const {
@@ -36,13 +30,6 @@ const MobileMenu = props => {
     onBlur,
     handleScrollToTop,
     username,
-    notifications,
-    userMetaData,
-    loadingNotifications,
-    notificationsPopoverVisible,
-    handleCloseNotificationsPopover,
-    getUserMetadata,
-    handleNotificationsPopoverVisibleChange,
     onLogout,
     disconnectPlatform,
     platformName,
@@ -59,90 +46,8 @@ const MobileMenu = props => {
     mobileMenuToggler();
   };
 
-  const handleMenuItemHide = () => {
-    mobileMenuToggler();
-  };
-
   const toggleModalBroker = () => {
     toggleModal('broker');
-  };
-
-  const buttonsMenuContainer = () => {
-    const lastSeenTimestamp = get(userMetaData, 'notifications_last_timestamp');
-    const notificationsCount = isUndefined(lastSeenTimestamp)
-      ? size(notifications)
-      : size(
-          filter(
-            notifications,
-            notification =>
-              lastSeenTimestamp < notification.timestamp &&
-              includes(PARSED_NOTIFICATIONS, notification.type),
-          ),
-        );
-    const displayBadge = notificationsCount > 0;
-    const notificationsCountDisplay = notificationsCount > 99 ? '99+' : notificationsCount;
-    return (
-      <div className="Topnav__menu-container">
-        <Menu selectedKeys={[]} className="Topnav__menu" mode="horizontal">
-          <Menu.Item className="Topnav__menu-item" key="write" onClick={handleMenuItemHide}>
-            <BTooltip
-              placement="bottom"
-              title={intl.formatMessage({ id: 'write_post', defaultMessage: 'Write post' })}
-              overlayClassName="Topnav__notifications-tooltip"
-              mouseEnterDelay={1}
-            >
-              <Link to="/editor" className="Topnav__link Topnav__link--action">
-                <i className="iconfont icon-write" />
-              </Link>
-            </BTooltip>
-          </Menu.Item>
-          <Menu.Item className="Topnav__menu-item" onClick={handleMenuItemHide}>
-            <Link to="/" className="Topnav__link Topnav__link--light Topnav__link--action">
-              <img className="feed" alt="feed" src={'/images/icons/ia-icon-feed.svg'} />
-            </Link>
-          </Menu.Item>
-          <Menu.Item className="Topnav__menu-item" key="hot">
-            <HotNews isMobile />
-          </Menu.Item>
-          <Menu.Item className="Topnav__menu-item" key="notifications">
-            <BTooltip
-              placement="bottom"
-              title={intl.formatMessage({ id: 'notifications', defaultMessage: 'Notifications' })}
-              overlayClassName="Topnav__notifications-tooltip"
-              mouseEnterDelay={1}
-            >
-              <PopoverContainer
-                placement="bottomRight"
-                trigger="click"
-                content={
-                  <Notifications
-                    notifications={notifications}
-                    onNotificationClick={handleCloseNotificationsPopover}
-                    st-card__chart
-                    currentAuthUsername={username}
-                    lastSeenTimestamp={lastSeenTimestamp}
-                    loadingNotifications={loadingNotifications}
-                    getUpdatedUserMetadata={getUserMetadata}
-                  />
-                }
-                visible={notificationsPopoverVisible}
-                onVisibleChange={handleNotificationsPopoverVisibleChange}
-                overlayClassName="Notifications__popover-overlay"
-                title={intl.formatMessage({ id: 'notifications', defaultMessage: 'Notifications' })}
-              >
-                <a className="Topnav__link Topnav__link--light Topnav__link--action">
-                  {displayBadge ? (
-                    <div className="Topnav__notifications-count">{notificationsCountDisplay}</div>
-                  ) : (
-                    <img alt="notifications" src={'/images/icons/ia-icon-bell.svg'} />
-                  )}
-                </a>
-              </PopoverContainer>
-            </BTooltip>
-          </Menu.Item>
-        </Menu>
-      </div>
-    );
   };
 
   return (
@@ -191,7 +96,7 @@ const MobileMenu = props => {
           </AutoComplete>
           <i className="iconfont icon-search" />
         </div>
-        {username ? buttonsMenuContainer() : <SignUp />}
+        {username ? <MenuButtons {...props} toggleMenu={mobileMenuToggler} /> : <SignUp />}
         <LoggedMenuMobile
           onLogout={onLogoutHandler}
           toggleMenu={mobileMenuToggler}
@@ -231,13 +136,6 @@ MobileMenu.propTypes = {
   onBlur: PropTypes.func.isRequired,
   handleScrollToTop: PropTypes.func.isRequired,
   username: PropTypes.string,
-  notifications: PropTypes.arrayOf(PropTypes.shape()),
-  userMetaData: PropTypes.shape(),
-  loadingNotifications: PropTypes.bool,
-  notificationsPopoverVisible: PropTypes.bool,
-  handleCloseNotificationsPopover: PropTypes.func.isRequired,
-  getUserMetadata: PropTypes.func.isRequired,
-  handleNotificationsPopoverVisibleChange: PropTypes.func.isRequired,
   onLogout: PropTypes.func.isRequired,
   disconnectPlatform: PropTypes.func.isRequired,
   platformName: PropTypes.string.isRequired,
@@ -249,19 +147,11 @@ MobileMenu.defaultProps = {
   username: '',
   dropdownOpen: false,
   searchBarValue: '',
-  notifications: [],
-  userMetaData: {},
-  loadingNotifications: false,
-  notificationsPopoverVisible: false,
 };
-
-const mapStateToProps = state => ({
-  isBrokerModalOpen: getModalIsOpenState(state, 'broker'),
-});
 
 const mapDispatchToProps = dispatch => ({
   onLogout: () => dispatch(logout()),
   disconnectPlatform: () => dispatch(disconnectBroker()),
 });
 
-export default injectIntl(connect(mapStateToProps, mapDispatchToProps)(MobileMenu));
+export default injectIntl(connect(null, mapDispatchToProps)(MobileMenu));
